nps-task/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bottle==0.12.20
bottle-websocket==0.2.9
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
Pillow==9.1.1
pymongo==3.12.3
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
sqlparse==0.2.4
urllib3==1.26.10
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
